+++ CHANGES.rst
@@ -3,6 +3,7 @@ Changelog (nionutils)
 
 0.3.21 (UNRELEASED)
 -------------------
+- Add geometry rectangle functions for intersect and union.
 - Add geometry functions to convert from int to float versions.
 
 0.3.20 (2020-01-28)